# backend/docs/openapi.yaml
components:
  schemas:
...
    Body_save_analysis_to_history_api_v1_history_analyses__analysis_id__save_post:
      properties:
        analysis_result:
          $ref: '#/components/schemas/GaitAnalysisResponse'
        video_info:
          anyOf:
          - type: object
          - type: 'null'
          title: Video Info
      required:
      - analysis_result
      title: Body_save_analysis_to_history_api_v1_history_analyses__analysis_id__save_post
      type: object
    GaitAnalysisResponse:
      description: 歩行分析レスポンス
      properties:
        analysis_id:
          description: 分析ID
          title: Analysis Id
          type: string
        analysis_settings:
          description: 分析設定
          title: Analysis Settings
          type: object
        gait_cycles:
          description: 検出された歩行周期
          items:
            $ref: '#/components/schemas/GaitCycle'
          title: Gait Cycles
          type: array
        gait_cycles_detected:
          description: 検出された歩行周期数
          title: Gait Cycles Detected
          type: integer
        joint_angles:
          anyOf:
          - $ref: '#/components/schemas/JointAngles'
          - type: 'null'
        overall_gait_score:
          description: 総合歩行スコア（0-100）
          title: Overall Gait Score
          type: number
        pose_detection_confidence:
          description: 姿勢検出信頼度
          title: Pose Detection Confidence
          type: number
        processing_time_seconds:
          description: 処理時間（秒）
          title: Processing Time Seconds
          type: number
        recommendations:
          description: 改善提案
          items:
            type: string
          title: Recommendations
          type: array
        spatiotemporal_params:
          anyOf:
          - $ref: '#/components/schemas/SpatiotemporalParameters'
          - type: 'null'
        status:
          description: 分析状態
          title: Status
          type: string
        symmetry_indices:
          anyOf:
          - $ref: '#/components/schemas/SymmetryIndices'
          - type: 'null'
        timestamp:
          description: 分析完了時刻
          format: date-time
          title: Timestamp
          type: string
        video_info:
          description: 動画情報
          title: Video Info
          type: object
        video_quality_score:
          description: 動画品質スコア（0-100）
          title: Video Quality Score
          type: number
      required:
      - analysis_id
      - status
      - processing_time_seconds
      - video_quality_score
      - pose_detection_confidence
      - gait_cycles_detected
      - overall_gait_score
      title: GaitAnalysisResponse
      type: object
    GaitCycle:
      description: 歩行周期データ
      properties:
        ankle_dorsiflexion_peak_deg:
          description: 足関節最大背屈角度
          title: Ankle Dorsiflexion Peak Deg
          type: number
        cycle_id:
          description: 周期ID
          title: Cycle Id
          type: string
        duration_seconds:
          description: 周期時間（秒）
          title: Duration Seconds
          type: number
        end_frame:
          description: 終了フレーム
          title: End Frame
          type: integer
        hip_flexion_peak_deg:
          description: 股関節最大屈曲角度
          title: Hip Flexion Peak Deg
          type: number
        knee_flexion_peak_deg:
          description: 膝関節最大屈曲角度
          title: Knee Flexion Peak Deg
          type: number
        phase_timings:
          additionalProperties:
            type: number
          description: フェーズタイミング
          title: Phase Timings
          type: object
        side:
          description: 左右（left/right）
          title: Side
          type: string
        stance_ratio:
          description: 立脚比率
          title: Stance Ratio
          type: number
        stance_time_seconds:
          description: 立脚時間（秒）
          title: Stance Time Seconds
          type: number
        start_frame:
          description: 開始フレーム
          title: Start Frame
          type: integer
        step_length_m:
          description: 歩幅（m）
          title: Step Length M
          type: number
        step_width_m:
          description: 歩隔（m）
          title: Step Width M
          type: number
        swing_time_seconds:
          description: 遊脚時間（秒）
          title: Swing Time Seconds
          type: number
      required:
      - cycle_id
      - start_frame
      - end_frame
      - duration_seconds
      - side
      - step_length_m
      - step_width_m
      - stance_time_seconds
      - swing_time_seconds
      - stance_ratio
      - hip_flexion_peak_deg
      - knee_flexion_peak_deg
      - ankle_dorsiflexion_peak_deg
      title: GaitCycle
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    JointAngles:
      description: 関節角度データ
      properties:
        ankle_dorsiflexion_deg:
          description: 足関節背屈角度時系列
          items:
            type: number
          title: Ankle Dorsiflexion Deg
          type: array
        ankle_plantarflexion_deg:
          description: 足関節底屈角度時系列
          items:
            type: number
          title: Ankle Plantarflexion Deg
          type: array
        ankle_toe_off_deg:
          default: 0.0
          description: トーオフ時足関節角度
          title: Ankle Toe Off Deg
          type: number
        frame_timestamps:
          description: フレームタイムスタンプ
          items:
            type: number
          title: Frame Timestamps
          type: array
        hip_abduction_deg:
          description: 股関節外転角度時系列
          items:
            type: number
          title: Hip Abduction Deg
          type: array
        hip_extension_deg:
          description: 股関節伸展角度時系列
          items:
            type: number
          title: Hip Extension Deg
          type: array
        hip_flexion_deg:
          description: 股関節屈曲角度時系列
          items:
            type: number
          title: Hip Flexion Deg
          type: array
        hip_rom_deg:
          default: 0.0
          description: 股関節可動域
          title: Hip Rom Deg
          type: number
        knee_flexion_deg:
          description: 膝関節屈曲角度時系列
          items:
            type: number
          title: Knee Flexion Deg
          type: array
        knee_heel_contact_deg:
          default: 0.0
          description: ヒールコンタクト時膝関節角度
          title: Knee Heel Contact Deg
          type: number
        knee_max_flexion_deg:
          default: 0.0
          description: 膝関節最大屈曲角度
          title: Knee Max Flexion Deg
          type: number
        pelvic_drop_deg:
          description: 骨盤傾斜角度時系列
          items:
            type: number
          title: Pelvic Drop Deg
          type: array
        pelvic_list_deg:
          description: 骨盤リスト（左右傾き）
          items:
            type: number
          title: Pelvic List Deg
          type: array
        trunk_rotation_deg:
          description: 体幹回旋角度時系列
          items:
            type: number
          title: Trunk Rotation Deg
          type: array
      title: JointAngles
      type: object
...
    SpatiotemporalParameters:
      description: 時空間パラメータ
      properties:
        cadence_steps_per_min:
          description: ケイデンス（歩/分）
          title: Cadence Steps Per Min
          type: integer
        double_support_time_s:
          description: 両脚支持時間（秒）
          title: Double Support Time S
          type: number
        gait_speed_ms:
          description: 歩行速度（m/s）
          title: Gait Speed Ms
          type: number
        left_stance_time_s:
          description: 左脚立脚時間（秒）
          title: Left Stance Time S
          type: number
        left_step_length_m:
          description: 左脚歩幅（m）
          title: Left Step Length M
          type: number
        right_stance_time_s:
          description: 右脚立脚時間（秒）
          title: Right Stance Time S
          type: number
        right_step_length_m:
          description: 右脚歩幅（m）
          title: Right Step Length M
          type: number
        stride_length_m:
          description: 歩幅（m）
          title: Stride Length M
          type: number
        stride_time_s:
          description: 歩行周期時間（秒）
          title: Stride Time S
          type: number
      required:
      - gait_speed_ms
      - cadence_steps_per_min
      - stride_length_m
      - stride_time_s
      - left_step_length_m
      - right_step_length_m
      - left_stance_time_s
      - right_stance_time_s
      - double_support_time_s
      title: SpatiotemporalParameters
      type: object
    SymmetryIndices:
      description: 対称性指標
      properties:
        ankle_dorsiflexion_symmetry_percent:
          description: 足関節背屈対称性（%）
          title: Ankle Dorsiflexion Symmetry Percent
          type: number
        hip_flexion_symmetry_percent:
          description: 股関節屈曲対称性（%）
          title: Hip Flexion Symmetry Percent
          type: number
        knee_flexion_symmetry_percent:
          description: 膝関節屈曲対称性（%）
          title: Knee Flexion Symmetry Percent
          type: number
        overall_symmetry_score:
          description: 総合対称性スコア（0-100）
          title: Overall Symmetry Score
          type: number
        stance_time_symmetry_percent:
          description: 立脚時間対称性（%）
          title: Stance Time Symmetry Percent
          type: number
        step_length_symmetry_percent:
          description: 歩幅対称性（%）
          title: Step Length Symmetry Percent
          type: number
        swing_time_symmetry_percent:
          description: 遊脚時間対称性（%）
          title: Swing Time Symmetry Percent
          type: number
      required:
      - step_length_symmetry_percent
      - stance_time_symmetry_percent
      - swing_time_symmetry_percent
      - hip_flexion_symmetry_percent
      - knee_flexion_symmetry_percent
      - ankle_dorsiflexion_symmetry_percent
      - overall_symmetry_score
      title: SymmetryIndices
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
    FirebaseAuth:
      flows:
        implicit:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          scopes:
            email: Email access
            openid: OpenID Connect
            profile: Profile access
      type: oauth2
info:
  contact:
    email: support@gaitanalysis.com
    name: Gait Analysis API Support
  description: "\n    ## Enhanced MediaPipe歩行分析API v2.0\n\n    スマートフォンで撮影した動画から歩行パターンを分析し、\n\
    \    理学療法士や医療従事者向けの詳細なレポートを生成する拡張APIです。\n\n    ### \U0001F680 新機能（v2.0 Enhanced）\n\
    \    - **30秒高速処理**: 撮影からレポート出力まで30秒以内\n    - **33ランドマーク活用**: MediaPipe Pose v0.10の全機能\n\
    \    - **代償動作検出**: 自動アラート機能付き\n    - **重心・バランス分析**: COM軌跡とバランス安定性評価\n    - **拡張出力**:\
    \ 正規化波形グラフ、スローモーション動画、カラーPDFレポート\n    - **iPad/iPhone最適化**: 60fps対応、横・後方撮影モード\n\
    \n    ### \U0001F4CA 詳細分析機能\n    - **スパイオテンポラル指標**: ピッチ、歩幅/ストライド長、立脚/遊脚時間、対称性指標\n\
    \    - **キネマティクス**: 股関節ROM、膝関節詳細角度、足関節背屈/底屈、骨盤・体幹回旋\n    - **代償動作アラート**: アウト/イントゥーイング、膝外反/内反、トレンデレンブルグ、ペルビックドロップ\n\
    \    - **バランス指標**: COM水平/垂直移動幅、重心投影線2Dマップ\n\n    ### \U0001F3AF 対象ユーザー\n    -\
    \ **理学療法士**: 詳細な歩行評価と治療計画立案\n    - **作業療法士**: 日常生活動作の評価\n    - **医師**: 診断支援と経過観察\n\
    \    - **研究者**: 歩行データの定量的解析\n    - **アスリート・トレーナー**: パフォーマンス向上\n\n    ### ⚙️ 技術仕様（Enhanced）\n\
    \    - **バックエンド**: FastAPI + Python 3.11+\n    - **AI処理**: MediaPipe Pose v0.10\
    \ + 最高精度モデル\n    - **並列処理**: マルチコア活用による高速化\n    - **データベース**: PostgreSQL + Redis\
    \ キャッシュ\n    - **認証**: Firebase Auth + JWT\n    - **監視**: リアルタイムパフォーマンス監視\n  \
    \  - **出力**: PDF/MP4/JSON/CSV 多形式対応\n    "
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  title: Enhanced MediaPipe Gait Analysis API
  version: 2.0.0-enhanced
openapi: 3.1.0
paths:
...
  /api/v1/history/analyses:
    get:
      description: "ユーザーの分析履歴を取得\n\nArgs:\n    user_id: ユーザーID\n    limit: 取得件数制限（1-100）\n\
        \    offset: オフセット\n    start_date: 開始日時\n    end_date: 終了日時\n    \nReturns:\n\
        \    JSONResponse: 分析履歴リスト"
      operationId: get_analysis_history_api_v1_history_analyses_get
      parameters:
      - description: ユーザーID
        in: query
        name: user_id
        required: true
        schema:
          description: ユーザーID
          title: User Id
          type: string
      - description: 取得件数制限
        in: query
        name: limit
        required: false
        schema:
          default: 20
          description: 取得件数制限
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - description: オフセット
        in: query
        name: offset
        required: false
        schema:
          default: 0
          description: オフセット
          minimum: 0
          title: Offset
          type: integer
      - description: 開始日時（ISO8601形式）
        in: query
        name: start_date
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 開始日時（ISO8601形式）
          title: Start Date
      - description: 終了日時（ISO8601形式）
        in: query
        name: end_date
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 終了日時（ISO8601形式）
          title: End Date
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Analysis History
      tags:
      - history
  /api/v1/history/analyses/{analysis_id}:
    delete:
      description: "分析結果を削除\n\nArgs:\n    analysis_id: 分析ID\n    user_id: ユーザーID\n\
        \    \nReturns:\n    JSONResponse: 削除結果"
      operationId: delete_analysis_api_v1_history_analyses__analysis_id__delete
      parameters:
      - in: path
        name: analysis_id
        required: true
        schema:
          title: Analysis Id
          type: string
      - description: ユーザーID
        in: query
        name: user_id
        required: true
        schema:
          description: ユーザーID
          title: User Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Analysis
      tags:
      - history
    get:
      description: "特定の分析結果の詳細を取得\n\nArgs:\n    analysis_id: 分析ID\n    user_id: ユーザーID\n\
        \    \nReturns:\n    JSONResponse: 分析結果詳細"
      operationId: get_analysis_detail_api_v1_history_analyses__analysis_id__get
      parameters:
      - in: path
        name: analysis_id
        required: true
        schema:
          title: Analysis Id
          type: string
      - description: ユーザーID
        in: query
        name: user_id
        required: true
        schema:
          description: ユーザーID
          title: User Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Analysis Detail
      tags:
      - history
  /api/v1/history/analyses/{analysis_id}/compare:
    get:
      description: "過去の分析結果と比較\n\nArgs:\n    analysis_id: 現在の分析ID\n    user_id: ユーザーID\n\
        \    compare_with: 比較対象の分析ID（未指定の場合は直前の分析）\n    \nReturns:\n    JSONResponse:\
        \ 比較結果"
      operationId: compare_with_previous_api_v1_history_analyses__analysis_id__compare_get
      parameters:
      - in: path
        name: analysis_id
        required: true
        schema:
          title: Analysis Id
          type: string
      - description: ユーザーID
        in: query
        name: user_id
        required: true
        schema:
          description: ユーザーID
          title: User Id
          type: string
      - description: 比較対象の分析ID
        in: query
        name: compare_with
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 比較対象の分析ID
          title: Compare With
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Compare With Previous
      tags:
      - history
  /api/v1/history/analyses/{analysis_id}/save:
    post:
      description: "分析結果を履歴に保存\n\nArgs:\n    analysis_id: 分析ID\n    user_id: ユーザーID\n\
        \    analysis_result: 分析結果データ\n    video_info: 動画情報\n    \nReturns:\n    JSONResponse:\
        \ 保存結果"
      operationId: save_analysis_to_history_api_v1_history_analyses__analysis_id__save_post
      parameters:
      - in: path
        name: analysis_id
        required: true
        schema:
          title: Analysis Id
          type: string
      - description: ユーザーID
        in: query
        name: user_id
        required: true
        schema:
          description: ユーザーID
          title: User Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_save_analysis_to_history_api_v1_history_analyses__analysis_id__save_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Save Analysis To History
      tags:
      - history
  /api/v1/history/statistics:
    get:
      description: "ユーザーの統計情報を取得\n\nArgs:\n    user_id: ユーザーID\n    \nReturns:\n \
        \   JSONResponse: 統計情報"
      operationId: get_user_statistics_api_v1_history_statistics_get
      parameters:
      - description: ユーザーID
        in: query
        name: user_id
        required: true
        schema:
          description: ユーザーID
          title: User Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get User Statistics
      tags:
      - history
...
servers:
...